refactor: extract CLI functions to separate file
